move AbstractID from runtime to core
