+++ flink-runtime/src/main/java/org/apache/flink/runtime/instance/Instance.java
@@ -26,7 +26,7 @@ import java.util.Queue;
 import java.util.Set;
 
 import akka.actor.ActorRef;
-import org.apache.flink.runtime.AbstractID;
+import org.apache.flink.util.AbstractID;
 import org.apache.flink.runtime.jobgraph.JobID;
 import org.apache.flink.runtime.jobmanager.scheduler.SlotAvailabilityListener;
 import org.apache.flink.runtime.jobmanager.scheduler.SlotSharingGroupAssignment;
