+++ flink-runtime/src/main/java/org/apache/flink/runtime/instance/InstanceID.java
@@ -18,7 +18,7 @@
 
 package org.apache.flink.runtime.instance;
 
-import org.apache.flink.runtime.AbstractID;
+import org.apache.flink.util.AbstractID;
 
 /**
  * Class for statistically unique instance IDs.
