+++ flink-runtime/src/main/java/org/apache/flink/runtime/instance/SimpleSlot.java
@@ -18,7 +18,7 @@
 
 package org.apache.flink.runtime.instance;
 
-import org.apache.flink.runtime.AbstractID;
+import org.apache.flink.util.AbstractID;
 import org.apache.flink.runtime.executiongraph.Execution;
 import org.apache.flink.runtime.jobgraph.JobID;
 import org.apache.flink.runtime.jobmanager.scheduler.Locality;
