+++ flink-runtime/src/main/java/org/apache/flink/runtime/instance/Slot.java
@@ -18,7 +18,7 @@
 
 package org.apache.flink.runtime.instance;
 
-import org.apache.flink.runtime.AbstractID;
+import org.apache.flink.util.AbstractID;
 import org.apache.flink.runtime.jobgraph.JobID;
 
 import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
