+++ flink-runtime/src/main/java/org/apache/flink/runtime/jobgraph/IntermediateResultPartitionID.java
@@ -19,7 +19,7 @@
 package org.apache.flink.runtime.jobgraph;
 
 import io.netty.buffer.ByteBuf;
-import org.apache.flink.runtime.AbstractID;
+import org.apache.flink.util.AbstractID;
 
 public class IntermediateResultPartitionID extends AbstractID {
 
@@ -34,6 +34,11 @@ public class IntermediateResultPartitionID extends AbstractID {
 
 	public IntermediateResultPartitionID(long lowerPart, long upperPart) {
 		super(lowerPart, upperPart);
+	}
+
+	public void writeTo(ByteBuf buf) {
+		buf.writeLong(this.lowerPart);
+		buf.writeLong(this.upperPart);
 	}
 
 	public static IntermediateResultPartitionID fromByteBuf(ByteBuf buf) {
